ZIO Module pattern 2.0
