+++ zio-aws-core/src/main/scala/io/github/vigoo/zioaws/core/config/ClientCustomization.scala
@@ -1,0 +1,20 @@
+package io.github.vigoo.zioaws.core.config
+
+import software.amazon.awssdk.awscore.client.builder.AwsClientBuilder
+
+trait ClientCustomization {
+  def customize[Client, Builder <: AwsClientBuilder[Builder, Client]](
+      builder: Builder
+  ): Builder
+}
+
+object ClientCustomization {
+
+  object None extends ClientCustomization {
+    override def customize[Client, Builder <: AwsClientBuilder[
+      Builder,
+      Client
+    ]](builder: Builder): Builder = builder
+  }
+
+}
