+++ zio-aws-core/src/main/scala/io/github/vigoo/zioaws/core/config/CommonAwsConfig.scala
@@ -1,0 +1,13 @@
+package io.github.vigoo.zioaws.core.config
+
+import software.amazon.awssdk.auth.credentials.AwsCredentialsProvider
+import software.amazon.awssdk.regions.Region
+
+import java.net.URI
+
+case class CommonAwsConfig(
+    region: Option[Region],
+    credentialsProvider: AwsCredentialsProvider,
+    endpointOverride: Option[URI],
+    commonClientConfig: Option[CommonClientConfig]
+)
