+++ zio-aws-core/src/main/scala/io/github/vigoo/zioaws/core/config/CommonClientConfig.scala
@@ -1,0 +1,10 @@
+package io.github.vigoo.zioaws.core.config
+
+import scala.concurrent.duration.Duration
+
+case class CommonClientConfig(
+    extraHeaders: Map[String, List[String]],
+    apiCallTimeout: Option[Duration],
+    apiCallAttemptTimeout: Option[Duration],
+    defaultProfileName: Option[String]
+)
