+++ zio-aws-core/src/main/scala/io/github/vigoo/zioaws/core/httpclient/ChannelOptions.scala
@@ -1,0 +1,3 @@
+package io.github.vigoo.zioaws.core.httpclient
+
+case class ChannelOptions(options: Vector[OptionValue[Any]])
