+++ zio-aws-core/src/main/scala/io/github/vigoo/zioaws/core/httpclient/OptionValue.scala
@@ -1,0 +1,5 @@
+package io.github.vigoo.zioaws.core.httpclient
+
+import java.net.SocketOption
+
+case class OptionValue[T](key: SocketOption[T], value: T)
