+++ zio-aws-core/src/main/scala/io/github/vigoo/zioaws/core/httpclient/ServiceHttpCapabilities.scala
@@ -1,0 +1,3 @@
+package io.github.vigoo.zioaws.core.httpclient
+
+case class ServiceHttpCapabilities(supportsHttp2: Boolean)
