+++ zio-aws-netty/src/main/scala/io/github/vigoo/zioaws/netty/Http2Config.scala
@@ -1,0 +1,9 @@
+package io.github.vigoo.zioaws.netty
+
+import zio.duration.Duration
+
+case class Http2Config(
+    maxStreams: Long,
+    initialWindowSize: Int,
+    healthCheckPingPeriod: Duration
+)
