+++ zio-aws-netty/src/main/scala/io/github/vigoo/zioaws/netty/HttpOrHttps.scala
@@ -1,0 +1,17 @@
+package io.github.vigoo.zioaws.netty
+
+sealed trait HttpOrHttps {
+  val asString: String
+}
+
+object HttpOrHttps {
+
+  case object Http extends HttpOrHttps {
+    override val asString: String = "http"
+  }
+
+  case object Https extends HttpOrHttps {
+    override val asString: String = "https"
+  }
+
+}
