+++ zio-aws-netty/src/main/scala/io/github/vigoo/zioaws/netty/NettyOptionValue.scala
@@ -1,0 +1,5 @@
+package io.github.vigoo.zioaws.netty
+
+import io.netty.channel.ChannelOption
+
+case class NettyOptionValue[T](key: ChannelOption[T], value: T)
