+++ zio-aws-netty/src/main/scala/io/github/vigoo/zioaws/netty/ProxyConfiguration.scala
@@ -1,0 +1,8 @@
+package io.github.vigoo.zioaws.netty
+
+case class ProxyConfiguration(
+    scheme: HttpOrHttps,
+    host: String,
+    port: Int,
+    nonProxyHosts: Set[String]
+)
